7-Gen_AI/Langchain_Campus_X/model/2.ChatModel/OpenAI.py
from langchain_openai import ChatOpenAI
from dotenv import load_dotenv

load_dotenv()

model = ChatOpenAI(model='gpt-4', temperature=0.9, max_completion_tokens=10)

result = model.invoke('What is the capital of India')
print(result)