7-Gen_AI/Langchain_Campus_X/requirements.txt
langchain
langchain-core
langchain-community

langchain-openai
openai

langchain-anthropic


langchain-google-genai
google-generativeai

langchain-huggingface
transformers
huggingface-hub

python-dotenv

numpy
pandas
scikit-learn
streamlit

PyPDFLoader